move handler to external package
